Introduce generic importer plugin view/controller
This commit introduces a base view and controller
(AbstractImporterView, AbstractImporterController) that can be used
to implement import plugins.
This base view features an import dialog with a common skeleton
for all import plugins. Derived classes can plug their specific
widgets in this common skeleton.

PCAPImport plugin was ported to this new architecture