Add support for import layer selection (see #243)
This commit provides the ability for the user to
select the network layer of which he wants to import the payload
while importing network messages. For example, the user is able
to import the ethernet payload (layer 2) or the IP payload (layer 3)
or the UDP/TCP payload (layer 4) of a given packet ans use it as a
message in Netzob.

NOTE: This commit introduces important changes in the way network
messages are handled in Netzob and *especially in the XML
representation of traces*. You may want to
review it thoroughly.

Three classes were created to represent nework messages:
- L2NetworkMessage (layer 2)
- L3NetworkMessage (layer 3)
- L4NetworkMessage (layer 4)
L4NetworkMessage inherits from L3Network which itself
inherits from L2NetworkMessage. Their respective
factories were also created.
*The common XML schema was modified to be able to store
these various types of messages*.

!!! IMPORTANT NOTE: Netzob is still able to import traces
following the old format BUT they will be automatically
converted to the new format when you save your project.
Please make sure your workspace does not contain valuable
data before testing this commit.

NetworkImport (network capture importer) and the OSPy importer
were modified to use L4NetworkMessage instead of the former
NetworkMessage class.

The PCAP import plugin is now able to import the payload of the layer
the user selected.

This commit also include some code style changes :
- Factories are now new-style classes
- Some comments were transformed into docstrings