Introduce a perspective system in main window
This commit add a support for perspectives in Netzob
main window.
Perspectives are composed of :
- A central panel
- Optional menu items to merge in Netzob main menu
- Optional toolbar items to merge in Netzob main
  toolbar

Netzob main menu and main toolbar, now have a common part
between all perspectives and a specific part, which is added
by the current perspective.

Perspectives should be composed of a View and a
Controller class. The controller class must have a "view" attribute
corresponding to its associated view class. The view must have
the following methods:
- getPanel() : Returns the widget which will be inserted
  in the central area of the main window
- getToolbarMenuUIDefinition() : Returns an XML UI definition
  (in GtkUIManager format) which will be merged in the
  GtkUIManager managing Netzob main toolbar
  and menu.
- getActionGroup() : Returns a Gtk.ActionGroup() holding
  actions using the the UI definition returned by the
  previous methods.

Toolbars and Menus of Netzob main window are now
managed by a GtkUIManager to allow easy merging of
toolbar items and menu items of the various
perspectives. Icons used by the toolbar items
are now retrieved from a GtkIconFactory (see
addNetzobIconsToDefaultFactory in NetzobWidgets) because
it is the only way to use custom icons with GtkUIManager.
This also allows icons to be moved from the ui/ directory to
the icons directory of static resources.

Former Grammar Inference and Simlulator perspectives as well as
the new vocabulary perspective made by Quentin Heyler were
ported to this new system.

NOTE: This commit was 'cherry-picked' in a new clean branch because
of issues in the branch history. When cherry-picking this commit it
introduced a modify/delete conflict on the files listed below.
They were added to the tree to ensure preserving modifications