Start a new modelization of variables. Data variables are modelized.
Add an AbstractVariable class, an AbstractLeafVariable class which depends on the previous one and a DataVariable which represents every leaf variable that has data.
(Formerly WordVariable, BinaryVariable ...)
The type of these variable is now managed by the Type class stored in the Type directory.
The main functions of the variable class are also modified and corrected in order to better follow the theoretical model.
They are also factorized.