Adds the toXML() and loadFromXML() functions to every Variable module.
This functions were in the former version of the code and have conserved their global performance.
toXML() export a variable in an xml tree.
loadFromXML() is a static function which create a variable from an XML definition.
Both functions comply to the same XML format.

Warning: this XML format has been modified since the former version of the code. Changes are the following:
 - The replacement of every former leaf variable (WordVariable, BinaryVariable ...) by a unique DataVariable has been ported to the xml format.
 - Each leaf variable has the following leaf (plus the former ones) containing values of attribute introduced in this version:
   - type;
   - minBits;
   - maxBits.