Remove useless computations when creating a project.
When creating a project, Netzob needs to verify if the name provided
by user is unique. To do so, it retrieves all the names of existing projects in the workspace
and compare them one by one with the new one.
Since this operation only requires names of other projects, we can use the
partial parsing of the xml to retrieve them.

In other word, we use method "getNameOfProjects" instead of "getProjects" from the workspace class.