Add the attribute node and functions findMotherNode(), removeChildByID(), editChildByID to variables.
The attribute node tells if a variable is a node (i.e. can have a child) or not. Every variable implements it.
RepeatVariable inherits now from AbstractNodeVariable and are node variable with only one child (the getChild() method returns self.children[0])
The function findMotherNode of the class AbstractVariable finds and returns the variable which has as child the given variable.
Functions removeChildByID() and editCHildByID() remove and edit a child of a node which would have the same ID has a given variable.