Implements a new variable management user interface.
Based on the previous model, this one allows the user to remove, modify and create subvariables for the variable of a given field. All these actions are granted by a simple interface with two dialog window. One for representing the tree representation of the variables. The other for editing/creating a variable according to the user's expectations.
The user interface is based on a set of glade files. This allows better reusability and adaptibility of the code.
Is uses the new definition of variables.
Field with no variable will automatically constructs a default one.