Recompute the default variable during message transfert.
If the user choose to transfer a message from a symbol to another, he modify the message structure of both symbols.
But if the fields of one (or both) symbol have still default variables, these variables are computed on the messages list of the symbol.
So if this list is updated, the default variable must be updated to.
That's what this commit do.