Introduce a treeview header for messages tables
The new treeview header widget, used in treeview
displaying messages of a given symbol, allows
to display the following properties regarding a field:
- name
- regex
- display format
The display format is hidden when the mouse pointer is
outside the message treeview. This is useful when more
than one message table is displayed in the right panel
of the vocabulary perspective and avoids eating up
to much vertical space to display information which is
not always useful.

It also allows the user to perform the following operations:
- changing the display format
- selecting or unselecting the field to perform operations
  on it (such as splitting it, concatenating fields etc...)
- collapsing the column (or expanding it if it is collapsed)
- sorting the column in ascending or descending order.

As regards technical details, it is possible to use
GtkTreeViewColumn.set_widget() to use a custom widget in a column
header, instead of a simple text label. The main issue is that
events are not propagated to this widget, which is a problem
if we want to make an interactive one. We therefore need to
propagate events by hand which requires hacking gtk a little.