Variable can be moved (relatively) in the variable tree view.
A new window is created for moving a variable to another position in the same node.
This allows the user to define more precisely and conveniently a data model based on variables.
A glade file has been created for this new interface.

Some small bugs have been resolved in the variable gui.
Some small enhancements have been added:
 - When changing the min size of a data variable, the max size modifies itself if it becomes inferior to the min size.
 - Same functionality for the number of iteration of a repeat variable.
 - The min size of a data variable fixes itself on the length of the given original value.