Make UDP usable in the netzob Simulator.
Before this commit TCP was usable but not UDP. An update had been applied to the TCP classes and not to the UDP ones. So the UDP-relative code was old and unfonctional.
This commit modifies, respectively, the ThreadedUDPServer and UDPConnectionHandler classes following the models of the ThreadedTCPServer and the TCPConnectionHandler classes in order to update these classes as the TCP ones were updated.