Corrects some mistakes in Type convertors.
bin2string used to call the bitarray tostring() function which can only traduced the 128 first hexadecimal values in string. This commit replaces this use by the tobytes() function which can translate all binary values.

bin2strhex used to not check if the binary it translates in strhex had characters or not. If not, it generates an error. So we now verify the size of the given binary, and if it is null we return a null string.