Modify the resources for the new relation variables.
The XSD that defined the format of saves is modified for the needs of the new relation variables. These variables have mandatory parameters to memorize that were not cover by the previous xsd model.
The VariableCreationView glade file is modified too, to allow the user to add these new relation variables in his netzob project through this same interface.