Modify all variable and the Field modules to update their import to Types and their management of the checked flag.
Types classes are now in the 'DataTypes' directory.
This commit implements also the two new variable classes in the loadFromXML function of AbstractVariable.
It modifies also the checked flag management for it to be conform to the very needs of the relation variable.
Now only nodes can decide if one of their child can be considered as checked or not.
Variables can no more decide by themselves.
And the checked flag is use in the read() and write() functions to protect variable for being edited twice which could lead to a relation variable with a value not related to the pointed variable's value.