Activate the quit operation.
When the user clicks on the close button (or the 'Quit' entry menu),
Netzob computes if the current project has pending modifications. If so,
it displays a dialog message to ask if the user wants to save the project before closing it.
Anyway, if the user answers yes or no, Netzob is closed.