Implements the new Variable glade file in the global VariableView and VariableController.
This commit changes the way the variable ui manages relation variable. They no more are kind of data variable with a pointed ID as original value.
Both types are now fully separated. An entire subpart of the VariableCreationView which elements are prefixed by ID or relation is dedicated to the relation variable.

This solves a lot of little bugs.