Introduce the Automatic Bug Reporter in Netzob.
If Netzob is started in devel mode (-d), any uncaught exceptions
are hooked and generates the execution of the ABR.

The ABR is a module which communicates with the bug tracker of Netzob
in order to create an issue for the current exception. The issue is created if
the user wants to and if the user has provided its API KEY.

In addition, when reporting a BUG, its SHA1 is computed and stored in a custom
field of the ticket.  This SHA1 objectives is to identify duplicated bugs
and in this case only register the user on it (adding a note).

Finaly, the provided API key is stored (if requested) in the user file (~/.netzob).