Add the mutateValue() function to all DataVariable type.
This function is called by DataVariable when they want to simply mutate their value instead of just send it or re-generate it.
It takes as parameter the value to mutate, a generation strategy list (just as the generateValue() function, and the strategies potentially given can be the same as for generation) rated of mutation, deletion and addition of characters.
It mutates accordingly to these rates a stringed version of the value, binarize it and give it back.
This function will be usefull whithin the next handling matrix for read and write access.