Implementation of the new variable treatment matrix.
Every variable's read and write value have been modified.
Both elementary treatments have been chopped down into many parts for more readability.

Relation variable are separated in two categories :
	direct relation variable that simply computes their value by the pointed variable's value. This is the case of variable that points to another symbol's variable or a left-placed variable.
	reverse relation variable that writes default value (retrived as if it was a direct relation variable) and binds itself to the pointed variable and wait for it to be modify and to notify it if so. At notification time the relation variable computes and updates the default values it has written in the processing token.

The processing token have been modified: re-organised and a new concept has been introduced :
ChoppedValue: contain the value (read or written) in a list, the elements of which corresponds to values that are entirely written by one leaf variable.
Each variable has a tokenChoppedIndexes list which contains the indexes of elements of the current processing token this variable is responsible for.
This is made for updating the value written by the reverse relation variables when they are notified.

This new matrix introduces a learn possibility to the node variable (aggregate, alternate and repeat) among the read access.
	aggregate variable can learn new children to comply to a given read value too much stuffed. New values are added under 0-1 repeat variables in order to become optional.
	alternate variable can learn new children if they did not manage to read a given part of value.
	repeat variable can adapt their minimum and maximum iterations number in order to comply with the given value and read the biggest part of it.
This introduces more flexibility in the simulation and a neuron network behavior.