Prevent current line to be selected when user toggle the checkbox from the left panel of the vocabulary view.
This change required to add a new method, which is in charge of
deciding if the current line should be selected or not
(symbol_list_selection_function). This function doesn't select the
current line if the symbol_list_set_selection flag is False. When user
toggle the checkbox, the flag's value is put to False.