Make the relation variables functional.
They used to not work.
They are constructed with a pointed ID (which is saved in the xml), and a symbol. This symbol allows the variable to gets its pointed Variable from the pointed ID, and to get back the current project or the root variable of the current symbol. It is used to randomize the pointed variable and find the position of the pointed variable relative to the pointing (relation) variable.

Moreover the computed relation variable has been adapt, it is now a mix between a direct relation variable and a data variable, it has a relation type which is linked to a data type.
So the toXML and loadFromXML functions of computed relations variables are rich and complexed, as the majority of functions.
The relation type tells what kind of relation type we use (size, checksum, ...) whereas the associated data type tells how the data of this relation (the size of the pointed block for instance) is stored.

The functions learn and compareFormat have been updated on the format of data variable's.