Allow the creation of SemiStochastic Transitions in the Grammar.
A SemiStochastic transition defines the transition from a starting state to an ending state.
This modification of the current state is triggered by the reception of an input symbol
and it triggers the emission of an output symbol.

Its a stochastic transition 'cause the output symbol is selected (randomly)
among a set of provided possible output symbol. Each output symbol
has a probability to be chosen and a reaction time.

A reaction represents the pause time before emiting the output symbol and is
defined in millisecond.