Prevent Netzob from starting if the current workspace is not valid
Before this patch, it was possible to start Netzob without specifying
a valid workspace. To occur, the user had to provide an invalid and yet existing workspace
directory twice. The second time, the XSD validation error didn't stop the application
from starting.

To fix this issue, the condition which verifies if the workspace exists and
is valid has been replaced by a while loop.