Organize the workspace selection flow using an MVC based pattern and fix IssueID #404.
The workspace selector execution flow is now based on an MVC which includes
a specific controller (netzob.Common.Controllers.WorkspaceSelectorController) and
its view (netzob.Common.Views.WorkspaceSelectorView). The view is based on a glade file
which can be found in resources/static/netzob/ui/workspaceSelectorDialog.glade.

In addition to the migration to an MVC pattern, it includes a bug fix and a small feature.
The error message is displayed on the dialog box when the loading of the specified workspace
has failed. Moreover, it won't stop from requesting the user a valid workspace unless he explicitly stops
the execution of the dialog by destroying it or by clicking on the cancel button (IssueID #404).

Change-Id: Ie61acb178a461fbb2de3cad19b706757c27d586e