Replace the SizeRelationType with two new types:
I splittes the SizeRelationType into two different types, depending on how their value is written eventually.
- BinaryRelationType: represent the size of a given field into a binary format.
- WordSizeRelationType: represent the size of a field into a string format.

These RelationTypes have been created with a new functionality in mind: implementation of a factor and an offset in their size computation. The final size value worth this now:
Final size = factor * (size in bits) + offset.
This gives users more flexibility on how the size is to be computed.