Update the getCells() method of Fields to reduce its complexity (and execution time) and returns a specific netzob list.
Method getCells() was recomputing alignments for every leaf fields. This is now changed and only one alignment per message is required
to returns the value of specified fields.

This commit also introduces the Class MessageTable which inherits from python built-in list. Its __repr__ method
is modified in order to show aligned messages in console and terminals.

Change-Id: I5885cdc6b5ea720655cf83beacc4b2822fd4531d