A new parser for Netzob: Major modification of its core.
This huge commit introduces a new parser for Netzob.
Sorry for the lack of details, but latter commits will be introduced to clean this working mess
and improves documentation and testing.
This new parser follows a traceback strategy to improve message alignment.
Computation time is increased but lots of algorithmic improvments can be provided to reduce this overhead.