Return code of the "setup.py test" command depends of test results
The objective of this commit is to ensure the interoperability of netzob unit-tests
with other tools such as travis-ci.
This commit ensures that the return code of Netzob unit-tests depends of tests results.
If at least one test fails, the "setup.py test" return code is set to False.
This is modification covers the test_command file used by setup.py to handle the
execution of netzob unit-tests.