feat(Raw/Alphabet): Introduces an alphabet in the definiton of a Raw Domain.
The alphabet of a Raw can be use to specify a subset of values the domain accepts.
This feature enables a finer grain on the values a field accepts. For example,
the following field definition can be use to produce base64-like values of 100 bytes.
It would be much more difficult to describe such domain with only Alternates and Raws.

>>> import string
>>> from netzob.all import *
>>> f = Field(Raw(nbBytes=100, alphabet=string.ascii_letters+string.digits+"+/"))